// generated/api/models/transaction.ts
import type { DecimalValue } from './decimalValue';
import type { Timestamp } from './timestamp';
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */
import type { TransactionType } from './transactionType';

export interface Transaction {
  amount?: DecimalValue;
  createdAt?: Timestamp;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  fromBankAccountId?: string | null;
  /** @nullable */
  toBankAccountId?: string | null;
  /** @nullable */
  transactionId?: string | null;
  type?: TransactionType;
}